// SMC.GPI.Client/projects/smc.gpi.inscricao/src/app/core/layout-processo/model/layout-navegacao.model.ts
export class LayoutNavegacaoModel {
  area: string = '';
  tokenPaginaAnteriorEncrypted: string = '';
  seqConfiguracaoEtapaPaginaAnteriorEncrypted: string = '';
  seqConfiguracaoEtapaEncrypted: string = '';
  seqGrupoOfertaEncrypted: string = '';
  seqInscricaoEncrypted: string = '';
  idioma: string = '';
  chaveTextoBotaoAnterior: string = '';
  inscricaoIniciada: boolean = false;
  chaveTextoBotaoProximo: string = '';
  seqInscricao: number = 0;
  uidProcesso: string = '';
}